WIP: Added boilerplate testing stuff.
