+++ lib/boilerplate/modes.js
@@ -16,9 +16,10 @@
  * - `config/webpack.config.js`   -> `PROJ/webpack.config.js`
  * - `templates/boilerplate/*`    -> `PROJ/*`: Generic files for all modes.
  *
- * TODO(exs): Refactor / rewrite this!!!
- * - `templates/{MODE}-slides/*`  -> `PROJ/src/*`: Presentation base
- * - `examples/cli-{MODE}/*`      -> `PROJ/src/*`: Example slide content
+ * As well as from published dependencies like:
+ *
+ * - `spectacle/examples/{js,md}`         -> `PROJ/src/*`
+ * - `spectacle-mdx-loader/examples/mdx`  -> `PROJ/src/*`
  *
  * ## Transformations
  *
@@ -34,7 +35,7 @@
  * - `PROJ/package.json`: Take CLI package.json dependencies and devDeps and
  *   transform them into appropriate ones for a standalone project. Adds various
  *   metadata and scripts, etc.
- * - `templates/{MODE}-slides/index.js`: We use the same regex replacement
+ * - `{DEP}/examples/{MODE}/index.js`: We use the same regex replacement
  *   scheme as the CLI does with the inject loader to hook up the example
  *   files in our examples. This results in our boilerplate output being
  *   nearly identical to our matching CLI examples.
@@ -42,7 +43,7 @@
 
 const path = require('path');
 const { copy, ensureDir, readFile, writeFile } = require('fs-extra');
-const { modes } = require('../templates/replacements');
+const { modes, getTmplDir } = require('../templates/replacements');
 
 // ----------------------------------------------------------------------------
 // Constants
@@ -192,40 +193,43 @@ const baseGenerate = async ({ name, description, dir, files }) => {
 // ----------------------------------------------------------------------------
 // Create a vanilla JS presentation.
 const js = async ({ name, description, dir }) => {
+  const tmplDir = getTmplDir('js');
   const files = {
-    '../templates/js-slides/index.html': 'src/index.html',
-    '../templates/js-slides/index.js': 'src/index.js'
+    [`${tmplDir}/index.html`]: 'src/index.html',
+    [`${tmplDir}/index.js`]: 'src/index.js'
   };
   await baseGenerate({ name, description, dir, files });
 };
 
 // Create an MDX-based presentation.
 const mdx = async ({ name, description, dir }) => {
+  const tmplDir = getTmplDir('mdx');
   const files = {
-    '../templates/mdx-slides/index.html': 'src/index.html',
-    '../templates/mdx-slides/index.js': {
+    [`${tmplDir}/index.html`]: 'src/index.html',
+    [`${tmplDir}/index.js`]: {
       dest: 'src/index.js',
       transform: transformIdx([
         { pattern: modes.mdx.re, replacement: modes.mdx.tmpl('./slides.mdx') }
       ])
     },
-    '../../examples/cli-mdx/slides.mdx': 'src/slides.mdx',
-    '../../examples/cli-mdx/test-component.js': 'src/test-component.js'
+    [`${tmplDir}/slides.mdx`]: 'src/slides.mdx',
+    [`${tmplDir}/test-component.js`]: 'src/test-component.js'
   };
   await baseGenerate({ name, description, dir, files });
 };
 
 // Create an MD-based presentation.
 const md = async ({ name, description, dir }) => {
+  const tmplDir = getTmplDir('mdx');
   const files = {
-    '../templates/md-slides/index.html': 'src/index.html',
-    '../templates/md-slides/index.js': {
+    [`${tmplDir}/index.html`]: 'src/index.html',
+    [`${tmplDir}/index.js`]: {
       dest: 'src/index.js',
       transform: transformIdx([
         { pattern: modes.md.re, replacement: modes.md.tmpl('./slides.md') }
       ])
     },
-    '../../examples/cli-md/slides.md': 'src/slides.md'
+    [`${tmplDir}/slides.md`]: 'src/slides.md'
   };
   await baseGenerate({ name, description, dir, files });
 };
