+++ lib/spectacle/actions.js
@@ -13,15 +13,9 @@ const {
   themeRe,
   themeTmpl,
   templateRe,
-  templateTmpl
+  templateTmpl,
+  getTmplDir
 } = require('../templates/replacements');
-const { findPkgRoot } = require('../util/file');
-
-// Table for finding published example templates;
-const TMPL_DIRS = {
-  md: () => path.resolve(findPkgRoot('spectacle'), `examples/md`),
-  mdx: () => path.resolve(findPkgRoot('spectacle-mdx-loader'), `examples/mdx`)
-};
 
 const webpackConfig = async ({
   mode = 'development',
@@ -35,10 +29,7 @@ const webpackConfig = async ({
   const srcMode = path.extname(srcFilePath).replace('.', '');
 
   // Templates, examples from dependencies.
-  const tmplDir = TMPL_DIRS[srcMode]();
-  if (!tmplDir) {
-    throw new Error(`Unabled to locate template directory for ${srcMode}`);
-  }
+  const tmplDir = getTmplDir(srcMode);
 
   // Webpack compiler + configuration.
   const entry = path.join(tmplDir, 'index.js');
