+++ lib/templates/replacements.js
@@ -1,6 +1,8 @@
 /**
  * Injection replacement patterns for all templates
  */
+const path = require('path');
+const { findPkgRoot } = require('../util/file');
 
 const jsRe = /\/\/ SPECTACLE_CLI_JS_START[\s\S]*?\/\/ SPECTACLE_CLI_JS_END/gm;
 const jsTmpl = srcFilePath => `
@@ -37,6 +39,22 @@ import template from '${templateFilePath}';
 // SPECTACLE_CLI_TEMPLATE_END
 `;
 
+// Table for finding published example templates;
+const TMPL_DIRS = {
+  js: () => path.resolve(findPkgRoot('spectacle'), `examples/js`),
+  md: () => path.resolve(findPkgRoot('spectacle'), `examples/md`),
+  mdx: () => path.resolve(findPkgRoot('spectacle-mdx-loader'), `examples/mdx`)
+};
+
+const getTmplDir = srcMode => {
+  const tmplDir = TMPL_DIRS[srcMode]();
+  if (!tmplDir) {
+    throw new Error(`Unabled to locate template directory for ${srcMode}`);
+  }
+
+  return tmplDir;
+};
+
 module.exports = {
   modes: {
     js: {
@@ -55,5 +73,6 @@ module.exports = {
   themeRe,
   themeTmpl,
   templateRe,
-  templateTmpl
+  templateTmpl,
+  getTmplDir
 };
