+++ package.json
@@ -31,12 +31,12 @@
     "boilerplate:start:md": "cd .examples-boilerplate/md && yarn start --port 6100",
     "boilerplate:start:mdx": "cd .examples-boilerplate/mdx && yarn start --port 6300",
     "examples:build:cli-md": "cd examples/cli-md && node ../../bin/spectacle/cli.js --title \"Spectacle CLI + MD Example\" --src ../../node_modules/spectacle/examples/md/slides.md -a build",
-    "examples:build:cli-mdx": "cd examples/cli-mdx && node ../../bin/spectacle/cli.js -l \"Spectacle CLI + MDX + Babel + Custom Theme Example\" --src ../../node_modules/spectacle-mdx-loader/examples/mdx/slides.mdx --action build --output dist",
-    "examples:build:cli-mdx-babel": "cd examples/cli-mdx-babel && node ../../bin/spectacle/cli.js -l \"Spectacle CLI + MDX + Babel + Spectacle Slide Example\" -q template.js -t theme.js --action build",
+    "examples:build:cli-mdx": "cd examples/cli-mdx && node ../../bin/spectacle/cli.js -l \"Spectacle CLI + MDX + Babel Example\" --src ../../node_modules/spectacle-mdx-loader/examples/mdx/slides.mdx --action build --output dist",
+    "examples:build:cli-mdx-babel": "cd examples/cli-mdx-babel && node ../../bin/spectacle/cli.js -l \"Spectacle CLI + MDX + Babel Example\" -q template.js -t theme.js --action build",
     "examples:build": "builder concurrent --buffer examples:build:cli-md examples:build:cli-mdx examples:build:cli-mdx-babel",
     "start:cli-md": "cd examples/cli-md && node ../../bin/spectacle/cli.js --title \"Spectacle CLI + MD Example\" --src ../../node_modules/spectacle/examples/md/slides.md --port 5100",
-    "start:cli-mdx": "cd examples/cli-mdx && node ../../bin/spectacle/cli.js -l \"Spectacle CLI + MDX + Custom Theme Example\" --src ../../node_modules/spectacle-mdx-loader/examples/mdx/slides.mdx -p 5000",
-    "start:cli-mdx-babel": "cd examples/cli-mdx-babel && node ../../bin/spectacle/cli.js -l \"Spectacle CLI + MDX + Babel + Spectacle Slide Example\" -q template.js -t theme.js -p 5001",
+    "start:cli-mdx": "cd examples/cli-mdx && node ../../bin/spectacle/cli.js -l \"Spectacle CLI + MDX Example\" --src ../../node_modules/spectacle-mdx-loader/examples/mdx/slides.mdx -p 5000",
+    "start:cli-mdx-babel": "cd examples/cli-mdx-babel && node ../../bin/spectacle/cli.js -l \"Spectacle CLI + MDX + Babel Example\" -q template.js -t theme.js -p 5001",
     "start:examples": "builder concurrent --log-level=info start:cli-md start:cli-mdx start:cli-mdx-babel",
     "start:boilerplate": "builder concurrent --log-level=info boilerplate:start:js boilerplate:start:md boilerplate:start:mdx",
     "start": "yarn start:cli-mdx",
